requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
cssbeautifier==1.15.3 ; python_version >= "3.10" and python_version < "4.0"
django-allauth==65.4.1 ; python_version >= "3.10" and python_version < "4.0"
django-browser-reload==1.18.0 ; python_version >= "3.10" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.10" and python_version < "4.0"
django-markdownx==4.0.7 ; python_version >= "3.10" and python_version < "4.0"
django-ranged-response==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-sendgrid-v5==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
django-simple-captcha==0.5.20 ; python_version >= "3.10" and python_version < "4.0"
django-storages==1.14.4 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.6 ; python_version >= "3.10" and python_version < "4.0"
djlint==1.36.4 ; python_version >= "3.10" and python_version < "4.0"
editorconfig==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.24.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-python-client==2.161.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.67.0 ; python_version >= "3.10" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
icalendar==5.0.13 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
jsbeautifier==1.15.3 ; python_version >= "3.10" and python_version < "4.0"
json5==0.10.0 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.26.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==5.29.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
python-avatars==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-http-client==3.3.7 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4.0"
sendgrid==6.11.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
starkbank-ecdsa==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
stripe==11.5.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.9.0 ; python_version >= "3.10" and python_version < "4.0"

# AI Assistant Dependencies
openai>=1.0.0
google-generativeai>=0.3.0
dotenv
nltk
tenacity

# Social Media Dependencies
tweepy==4.15.0 ; python_version >= "3.10" and python_version < "4.0"
